requirements.txt
flask
sentence-transformers
datasets
torch
